Extract Autocomplete from PreferenceForm into CuisineAutocomplete component
